Implementation of a pipeline driven image analysis workflow
Writing pipeline logic for analysing images. The pipeline is configured through a human readable yaml file and defines a linear work flow. Each step in the pipeline is encapsulated in each python class, and generates a desired result, which can be consumed by subsequent processes. The pipeline performs a large set of data validation steps, like
- checking whether the pipeline has all required inputs and delivers all required data sets
- whether each input to a given process has the right data type